requirements.txt
playwright>=1.45,<1.47
PyYAML>=6.0
rich>=13.7
python-dotenv>=1.0
pyautogui>=0.9.54
opencv-python>=4.8
Pillow>=10.0
pyscreeze>=0.1.28
